% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{build_app}
\alias{build_app}
\title{Opens the shiny app in a web browser}
\usage{
build_app()
}
\value{
A Shiny app object
}
\description{
Opens the shiny app in a web browser
}
\examples{
app <- build_app()
if (interactive()) shiny::runApp(app)
}
