% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{.build_server}
\alias{.build_server}
\title{Create server block for Shiny}
\usage{
.build_server()
}
\value{
Server for \code{\link[shiny:shinyApp]{shiny::shinyApp()}}
}
\description{
Create server block for Shiny
}
\author{
Seung J. Kim
}
\keyword{internal}
