% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{.build_ui}
\alias{.build_ui}
\title{Create UI block for Shiny}
\usage{
.build_ui()
}
\value{
UI for \code{\link[shiny:shinyApp]{shiny::shinyApp()}}
}
\description{
Create UI block for Shiny
}
\author{
Seung J. Kim
}
\keyword{internal}
