% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_heatmap_nav_panel.R
\name{.interfaceHeatmapNavPanel}
\alias{.interfaceHeatmapNavPanel}
\title{Create the "Heatmap" nav panel}
\usage{
.interfaceHeatmapNavPanel()
}
\value{
\code{\link[bslib:nav-items]{bslib::nav_panel()}}
}
\description{
Create the "Heatmap" nav panel
}
\author{
Seung J. Kim
}
\keyword{internal}
