% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_intro_nav_panel.R
\name{.interfaceIntroNavPanel}
\alias{.interfaceIntroNavPanel}
\title{Create the "QC" nav panel}
\usage{
.interfaceIntroNavPanel()
}
\value{
\code{\link[bslib:nav-items]{bslib::nav_panel()}}
}
\description{
Create the "QC" nav panel
}
\author{
Seung J. Kim
}
\keyword{internal}
