% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_pca_nav_panel.R
\name{.interfacePcaNavPanel}
\alias{.interfacePcaNavPanel}
\title{Create the "PCA" nav panel}
\usage{
.interfacePcaNavPanel()
}
\value{
\code{\link[bslib:nav-items]{bslib::nav_panel()}}
}
\description{
Create the "PCA" nav panel
}
\author{
Seung J. Kim
}
\keyword{internal}
