% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_qc_nav_panel.R
\name{.interfaceQcNavPanel}
\alias{.interfaceQcNavPanel}
\title{Create the "QC" nav panel}
\usage{
.interfaceQcNavPanel()
}
\value{
\code{\link[bslib:nav-items]{bslib::nav_panel()}}
}
\description{
Create the "QC" nav panel
}
\author{
Seung J. Kim
}
\keyword{internal}
