% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sidebar.R
\name{.interfaceSidebar}
\alias{.interfaceSidebar}
\title{Creates the "sidebar" UI element}
\usage{
.interfaceSidebar()
}
\value{
\code{\link[bslib:sidebar]{bslib::sidebar()}}
}
\description{
Creates the "sidebar" UI element
}
\author{
Seung J. Kim
}
\keyword{internal}
