% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_table_nav_panel.R
\name{.interfaceTableNavPanel}
\alias{.interfaceTableNavPanel}
\title{Create the "Table" nav panel}
\usage{
.interfaceTableNavPanel()
}
\value{
\code{\link[bslib:nav-items]{bslib::nav_panel()}}
}
\description{
Create the "Table" nav panel
}
\author{
Seung J. Kim
}
\keyword{internal}
