% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_volcano_nav_panel.R
\name{.interfaceVolcanoNavPanel}
\alias{.interfaceVolcanoNavPanel}
\title{Create the "Volcano" nav panel}
\usage{
.interfaceVolcanoNavPanel(output)
}
\value{
\code{\link[bslib:nav-items]{bslib::nav_panel()}}
}
\description{
Create the "Volcano" nav panel
}
\author{
Seung J. Kim
}
\keyword{internal}
