% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Helper function that exposes .png assets to the Shiny package}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{libname}

\item{pkgname}{pkgname}
}
\value{
\code{\link[=shiny:addResourcePath]{shiny:addResourcePath()}}
}
\description{
Helper function that exposes .png assets to the Shiny package
}
\keyword{internal}
