% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_PCA_customization.R
\name{.PCAcustomization}
\alias{.PCAcustomization}
\title{Create a group of radio buttons for picking shapes for PCA plots}
\usage{
.PCAcustomization()
}
\value{
A list of \code{\link[shinyWidgets:radioGroupButtons]{shinyWidgets::radioGroupButtons()}} and
\code{\link[shiny:textInput]{shiny::textInput()}} pairs
}
\description{
Create a group of radio buttons for picking shapes for PCA plots
}
\author{
Seung J. Kim
}
\keyword{internal}
