% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_PCA_function.R
\name{.PCAFunction}
\alias{.PCAFunction}
\title{PCA plotting function with ggplot2}
\usage{
.PCAFunction(
  spe,
  precomputed,
  colourShapeBy,
  selectedVar,
  ROIshapes,
  ROIcolours
)
}
\arguments{
\item{spe}{A \link{SpatialExperiment} output from \code{\link[standR:readGeoMx]{standR::readGeoMx()}}}

\item{precomputed}{Output from \code{\link[SingleCellExperiment:reducedDims]{SingleCellExperiment::reducedDim()}}}

\item{colourShapeBy}{From input$selected_types}

\item{selectedVar}{Either "Type" or input$selected_batch}

\item{ROIshapes}{User input shapes from .PCA_customization() function OR
PCA_customization_batch()}

\item{ROIcolours}{User input colours from .PCA_customization() function OR
PCA_customization_batch()}
}
\value{
A \link{ggplot2} object
}
\description{
PCA plotting function with ggplot2
}
\author{
Seung J. Kim
}
\keyword{internal}
