% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_volcano_function.R
\name{.volcanoFunction}
\alias{.volcanoFunction}
\title{Volcano plot plotting function with ggplot2}
\usage{
.volcanoFunction(
  volcano,
  delabSize,
  maxOverlap,
  title,
  logFCcutoff,
  PvalCutoff,
  DnCol,
  notDEcol,
  UpCol
)
}
\arguments{
\item{volcano}{a data.frame}

\item{delabSize}{from input$delabSize}

\item{maxOverlap}{from input$maxOverlap}

\item{title}{Contrasts title}

\item{logFCcutoff}{from input$logFCcutoff}

\item{PvalCutoff}{from input$PvalCutoff}

\item{DnCol}{from input$DnCol. Determines the colour of downregulated genes.}

\item{notDEcol}{from input$notDEcol.}

\item{UpCol}{from input$UpCol. Determines the colour of upregulated genes.}
}
\value{
A \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} object
}
\description{
Volcano plot plotting function with ggplot2
}
\author{
Seung J. Kim
}
\keyword{internal}
