% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ShinyDSP}
\alias{ShinyDSP}
\title{Opens the shiny app in a web browser}
\usage{
ShinyDSP()
}
\value{
A \code{\link[shiny:shinyApp]{shiny::shinyApp()}} object
}
\description{
Opens the shiny app in a web browser
}
\examples{
library(ShinyDSP)
if (interactive()) shiny::runApp(ShinyDSP)
}
\author{
Seung J. Kim
}
