% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDSP-package.R
\docType{package}
\name{shinyDSP-package}
\alias{shinyDSP-package}
\title{shinyDSP: A Shiny App For Visualizing Nanostring GeoMx DSP Data}
\description{
This package is a Shiny app for interactively analyzing and visualizing Nanostring GeoMX Whole Transcriptome Atlas data. Users have the option of exploring a sample data to explore this app's functionality. Regions of interest (ROIs) can be filtered based on any user-provided metadata. Upon taking two or more groups of interest, all pairwise and ANOVA-like testing are automatically performed. Available ouputs include PCA, Volcano plots, tables and heatmaps. Aesthetics of each output are highly customizable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kimsjune.github.com/shinyDSP}
  \item \url{http://joonkim.ca/shinyDSP/}
  \item Report bugs at \url{https://kimsjune.github.com/ShinyDSP/issues}
}

}
\author{
\strong{Maintainer}: Seung J. Kim \email{skim823@uwo.ca} (\href{https://orcid.org/0000-0001-5263-0758}{ORCID})

Other contributors:
\itemize{
  \item Marco Mura \email{marco.mura@lhsc.on.ca} [funder]
}

}
\keyword{internal}
