% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDSP-package.R
\docType{package}
\name{shinyDSP-package}
\alias{shinyDSP-package}
\title{shinyDSP: A Shiny App For Visualizing Nanostring GeoMx DSP Data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kimsjune.github.com/ShinyDSP}
  \item \url{http://joonkim.ca/shinyDSP/}
  \item Report bugs at \url{https://kimsjune.github.com/ShinyDSP/issues}
}

}
\author{
\strong{Maintainer}: Seung J. Kim \email{seung.kim@lhsc.on.ca} (\href{https://orcid.org/0000-0001-5263-0758}{ORCID})

Other contributors:
\itemize{
  \item Angwara Nilanont \email{angwara.nilanont@lhsc.on.ca} (\href{https://orcid.org/0000-0003-4468-9172}{ORCID}) [contributor]
  \item Marco Mura \email{marco.mura@lhsc.on.ca} [funder]
}

}
\keyword{internal}
