% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyDSP-package.R
\docType{package}
\name{ShinyDSP-package}
\alias{ShinyDSP}
\alias{ShinyDSP-package}
\title{ShinyDSP: A Shiny App For Visualizing Nanostring GeoMx DSP Data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kimsjune.github.com/ShinyDSP}
  \item Report bugs at \url{https://kimsjune.github.com/ShinyDSP/issues}
}

}
\author{
\strong{Maintainer}: Seung Kim \email{seung.kim@lhsc.on.ca} (\href{https://orcid.org/0000-0001-5263-0758}{ORCID})

}
\keyword{internal}
